+++ src/redux/actions/set-search-result.js
